Removed old Note fragment with presenter etc.
